Save live suffix into preset; allow decimal bitrates (v1.0007)
- Save/Save As now bake whatever's typed in the Output section's Suffix
  field into the preset's own outputSuffix, the same role Clearcast
  MP4's built-in "_CC" plays. Falls back to the preset's existing
  suffix when the field is left blank, so saving without touching it
  doesn't erase anything. The live field clears afterward since the
  value is now the preset's own default, shown via placeholder.
- Bitrate field validation now accepts decimals (Double) instead of
  only whole numbers (Int) - fractional Mbps like 12.5 was already
  valid ffmpeg input and already flowed through the real encode/estimate
  paths unchanged; this was the one leftover Int-only validation
  blocking the Encode button.

+++ Transcoda/ContentView.swift
@@ -34,10 +34,13 @@ struct ContentView: View {
                 guard let v = Double(maxSize), v > 0 else { return false }
             } else {
                 // Blank bitrate is fine — PresetConfig falls back to the default.
-                // Anything entered must still be a valid positive number, though.
+                // Anything entered must still be a valid positive number, though
+                // (Double, not Int — fractional Mbps like 12.5 is valid ffmpeg
+                // input and already flows through the actual encode/estimate
+                // paths unchanged; this was the only place still Int-only).
                 let trimmed = settings.bitrateMbps.trimmingCharacters(in: .whitespaces)
                 if !trimmed.isEmpty {
-                    guard let v = Int(trimmed), v > 0 else { return false }
+                    guard let v = Double(trimmed), v > 0 else { return false }
                 }
             }
         }
